[Core] Move Vector2i and Rect2i out of math_2d.h
